add lombok

+++ ego-parent/ego-service-impl/src/main/java/com/ego/mapper/TbContentCategoryMapper.java
@@ -2,11 +2,12 @@ package com.ego.mapper;
 
 import com.ego.pojo.TbContentCategory;
 import com.ego.pojo.TbContentCategoryExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface TbContentCategoryMapper {
-    int countByExample(TbContentCategoryExample example);
+    long countByExample(TbContentCategoryExample example);
 
     int deleteByExample(TbContentCategoryExample example);
