+++ ego-parent/ego-service-impl/src/main/java/com/ego/mapper/TbItemMapper.java
@@ -2,11 +2,12 @@ package com.ego.mapper;
 
 import com.ego.pojo.TbItem;
 import com.ego.pojo.TbItemExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface TbItemMapper {
-    int countByExample(TbItemExample example);
+    long countByExample(TbItemExample example);
 
     int deleteByExample(TbItemExample example);
