+++ ego-parent/ego-service-impl/src/main/java/com/ego/mapper/TbOrderItemMapper.java
@@ -2,11 +2,12 @@ package com.ego.mapper;
 
 import com.ego.pojo.TbOrderItem;
 import com.ego.pojo.TbOrderItemExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface TbOrderItemMapper {
-    int countByExample(TbOrderItemExample example);
+    long countByExample(TbOrderItemExample example);
 
     int deleteByExample(TbOrderItemExample example);
