+++ ego-parent/ego-service-impl/src/main/java/com/ego/mapper/TbOrderMapper.java
@@ -2,11 +2,12 @@ package com.ego.mapper;
 
 import com.ego.pojo.TbOrder;
 import com.ego.pojo.TbOrderExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface TbOrderMapper {
-    int countByExample(TbOrderExample example);
+    long countByExample(TbOrderExample example);
 
     int deleteByExample(TbOrderExample example);
