+++ ego-parent/ego-service-impl/src/main/java/com/ego/mapper/TbUserMapper.java
@@ -2,11 +2,12 @@ package com.ego.mapper;
 
 import com.ego.pojo.TbUser;
 import com.ego.pojo.TbUserExample;
-import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
+import java.util.List;
+
 public interface TbUserMapper {
-    int countByExample(TbUserExample example);
+    long countByExample(TbUserExample example);
 
     int deleteByExample(TbUserExample example);
